net: change the type of id in icmpv[6]_findconn to uint16_t
since the id field in ICMP[6] is 16bits length

+++ net/icmp/icmp_conn.c
@@ -269,7 +269,7 @@ FAR struct icmp_conn_s *icmp_nextconn(FAR struct icmp_conn_s *conn)
  *
  ****************************************************************************/
 
-FAR struct icmp_conn_s *icmp_findconn(FAR struct net_driver_s *dev, uint8_t id)
+FAR struct icmp_conn_s *icmp_findconn(FAR struct net_driver_s *dev, uint16_t id)
 {
   FAR struct icmp_conn_s *conn;
